requirements.txt
# requirements.txt
# 基础依赖
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
python-docx>=0.8.11
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
jinja2>=3.0.0

# LangChain相关依赖
langchain>=0.0.267
langchain-openai>=0.0.2
openai>=0.27.0

# 数据处理相关
scipy>=1.7.0
scikit-learn>=1.0.0
statsmodels>=0.13.0
openpyxl>=3.0.9  # 用于Excel文件处理
python-dotenv>=0.19.0  # 用于环境变量管理